Address review feedback: accessibility, pointer-events, extract TipsVersionInit
- Add aria-labels to lightbulb and Learn more buttons (a11y)
- Add pointer-events-none to tip bar wrapper, pointer-events-auto on interactive children
- Throttle mousemove idle timer resets (10s debounce)
- Add Escape key handler to dismiss TipSpot popovers
- Extract TipsVersionInit from App.tsx into tips/tips-version-init.tsx
