+++ apps/web/src/components/app/agents-view.tsx
@@ -630,7 +630,7 @@ export function AgentsView({
                   }
                 />
                 {!isMobile ? (
-                  <div className="absolute inset-x-0 bottom-0 z-10 h-14">
+                  <div className="pointer-events-none absolute inset-x-0 bottom-0 z-10 h-14">
                     <AmbientTipBar />
                   </div>
                 ) : null}
